NLP: improve language detection
Check the ratio of roman letters in the whole content, cancel language detection if they are not 90% of the content.

Check that we have at least 10% of stopwords in the content, matching the detected language, else return None for the result is not significant.

This explicitely disables support for non-latin alphabets (Greek, Cyrillic, Japanese/Chinese).